added login register UI
