+++ pom.xml
@@ -58,6 +58,12 @@
             <version>42.2.5</version> <!-- Use the latest version -->
         </dependency>
 
+        <dependency>
+            <groupId>jline</groupId>
+            <artifactId>jline</artifactId>
+            <version>3.0.0.M1</version>
+        </dependency>
+
         <!-- live reload on code changes (for Dev Env only) -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
