--- src/main/java/play/blackjack/utils/PrintDashes.java
@@ -1,13 +1,0 @@
-package play.blackjack.utils;
-
-public class PrintDashes {
-    public static void printDashes(int n) {
-        for (int i = 0; i < n; i++) {
-            System.out.print("-");
-        }
-        System.out.println();
-    }
-    public static void printDashes() {
-        printDashes(50);
-    }
-}
