+++ src/main/java/play/blackjack/utils/UserInterface.java
@@ -1,0 +1,26 @@
+package play.blackjack.utils;
+
+public class UserInterface {
+    public static void printDashes(int n) {
+        for (int i = 0; i < n; i++) {
+            System.out.print("-");
+        }
+        System.out.println();
+    }
+    public static void printDashes() {
+        printDashes(43);
+    }
+
+    // using cmd and terminal logic doesn't work on ALL OS.
+    public static void clearScreen() {
+        for (int i = 0; i < 5; i++) {
+            System.out.println(
+                    "\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n" +
+                    "\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n" +
+                    "\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n" +
+                    "\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n");
+        }
+    }
+
+
+}
